+++ shaders/landscape.vert
@@ -1,12 +1,12 @@
 #version 460
 #extension GL_ARB_separate_shader_objects : enable
 
-layout (location=0) in vec2 coordIn;
+layout (location=0) in ivec2 coordIn;
 
 layout (location=0) uniform mat4 m_transform;
 layout (location=1) uniform mat4 m_transform1;
 
-layout (location=2) uniform vec4 levelShift;
+layout (location=2) uniform ivec4 levelShift;
 layout (location=3) uniform float heightScale;
 
 layout (location=7) uniform float step;
@@ -23,8 +23,8 @@ layout (location=2) out float brightness;
 
 void main(void) {
     vec3 coord3d = vec3(float(coordIn.x) * step, float(coordIn.y) * step, 0.0);
-    uint index = (int(coordIn.y + levelShift.y) % width) * width
-               + (int(coordIn.x + levelShift.x) % width);
+    uint index = ((coordIn.y + levelShift.y) % width) * width
+               + ((coordIn.x + levelShift.x) % width);
     vec3 coordf = vec3(coord3d.x, coord3d.y, float(heights[index]) * heightScale);
     vec4 coord = m_transform * m_transform1 * vec4(coordf, 1.0);
     gl_Position = coord;
